Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/url-attribute-expected.txt

Both .URL and .url attributes should work (for compatibility reasons).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.url is "ws://127.0.0.1:8880/"
PASS ws.URL is "ws://127.0.0.1:8880/"
PASS ws.URL === ws.url is true
PASS successfullyParsed is true

TEST COMPLETE
